--- src/components/CartController/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './CartController';
