+++ src/components/Header/Header.style.ts
@@ -11,28 +11,3 @@ export const HeaderWrapper = styled.div`
   align-items: center;
   height: 80px;
 `;
-
-export const CartWrapper = styled.button`
-  display: flex;
-  color: ${({ theme }) => theme.lightColor};
-`;
-
-export const CartTitle = styled.div`
-  font-size: 24px;
-`;
-
-export const CartCountWrapper = styled.div`
-  border-radius: 50%;
-  background-color: #04c09e;
-  width: 26px;
-  height: 26px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-left: 8px;
-`;
-
-export const CartCount = styled.div`
-  font-size: 16px;
-  color: #fff;
-`;
